refactor filter and exception
