+++ async_casbin_sqlmodel_adapter/adapter.py
@@ -11,24 +11,10 @@ from sqlalchemy.orm import sessionmaker
 from sqlalchemy.sql.dml import Delete
 from sqlmodel import SQLModel, delete, or_, select
 from sqlmodel.sql.expression import SelectOfScalar
+from .exception import AdapterException
+from .filter import Filter
 
 
-class AdapterException(Exception):
-    """AdapterException"""
-
-
-class Filter:  # pylint: disable=too-few-public-methods
-    """
-    Filter class for SQLModel-based Casbin adapter.
-    """
-
-    ptype: list[str] = []
-    v0: list[str] = []
-    v1: list[str] = []
-    v2: list[str] = []
-    v3: list[str] = []
-    v4: list[str] = []
-    v5: list[str] = []
 
 
 class Adapter(BaseAdapter, UpdateAdapter):
