+++ async_casbin_sqlmodel_adapter/exception.py
@@ -1,0 +1,2 @@
+class AdapterException(Exception):
+    """AdapterException"""
