+++ async_casbin_sqlmodel_adapter/filter.py
@@ -1,0 +1,12 @@
+class Filter:  # pylint: disable=too-few-public-methods
+    """
+    Filter class for SQLModel-based Casbin adapter.
+    """
+
+    ptype: list[str] = []
+    v0: list[str] = []
+    v1: list[str] = []
+    v2: list[str] = []
+    v3: list[str] = []
+    v4: list[str] = []
+    v5: list[str] = []
